… by avoiding adding dummy dependencies if the project's python constraint does not allow any version compatible with the marker of the dependency

follow-up of #4695

Requires: python-poetry/poetry-core#347 (poetry-core > 1.1.0a7)

I did some measurements from issues solved by #4695 with and without this PR. I always measured the second resolution so that a complete cache before each measurement can be assumed. The differences in resolution time may be greater without a filled cache. (The resulting lock file is the same with/without this PR.)

|`pyproject.toml` from ...| without PR | with PR |
|---|---|---|
|#3367|Complete version solving took 0.401 seconds with 6 overrides|Complete version solving took 0.370 seconds with 4 overrides|
|#4670|Complete version solving took 0.733 seconds with 24 overrides|Complete version solving took 0.590 seconds with 16 overrides|
|#4870|Complete version solving took 31.120 seconds with 57 overrides|Complete version solving took 27.163 seconds with 52 overrides|